Setup exec library and updated Makefile
