+++ src/psil_exec_types.cpp
@@ -1,0 +1,28 @@
+/**
+   psil_exec_types.cpp
+   PSIL Execution Library
+   Type checks and
+   Type conversion implementations
+   @author Sinclair Gurny
+   @version 0.5
+   July 2019
+*/
+
+#include "psil_exec.h"
+
+namespace psil_exec {
+
+  // TODO
+
+  // Identity predicates
+  void psil_isbool( stack_ptr & s, token_ptr & node ) {
+    // call check type?
+  }
+  void psil_isnum( stack_ptr & s, token_ptr & node ) {}
+  void psil_ischar( stack_ptr & s, token_ptr & node ) {}
+  void psil_issymbol( stack_ptr & s, token_ptr & node ) {}
+  void psil_isproc( stack_ptr & s, token_ptr & node ) {}
+  void psil_islist( stack_ptr & s, token_ptr & node ) {}
+
+
+}
